mathias-8/Mathias-task8-task-202411060853/gen_test/GameTest.java
package test;

import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.*;

public class GameTest {

    @Before
    public void setUp() {
        Game.state = null;
    }

    @Test
    public void testGenerateRoomsFromFileValid() {
        // Assume the file "rooms.txt" is set up properly for testing
        Game.generateRoomsFromFile();
        assertNotNull(Game.state);
    }

    @Test(expected = IllegalStateException.class)
    public void testGenerateRoomsFromFileInvalid() {
        // Point to a non-existent file or incorrectly formatted file
        Game.generateRoomsFromFile();
    }

    @Test
    public void testGetCommand() {
        // Simulate user input
        String input = " go east ";
        InputStream in = new ByteArrayInputStream(input.getBytes());
        System.setIn(in);

        String command = Game.getCommand();
        assertEquals("go east", command);
    }

}
```

Please create separate files for each test class according to Java naming conventions: `RoomTest.java`, `GameStateTest.java`, `CommandParserTest.java`, and `GameTest.java`. Each test class contains a variety of tests targeting different functionalities of the game, simulating user commands and examining room changes. Adjust the setup of the `GameTest` to work with an actual rooms configuration.